+++ cxl/vm.sh
@@ -137,7 +137,7 @@ custom_cxl() {
 if [[ ! -z ${CUSTOM} ]]; then
 	custom_cxl
 else
-	cxlargs+=( --cxl device=cxl-pmem-4way )
+	cxlargs+=( --cxl device=cxl-vmem-4way-dc )
 fi
 
 [[ -z ${NOCXL} ]] && qargs+=( ${cxlargs[@]} )